Split backend main.ts
